refactor(kernel): rename the MegaMoE kernel to sm100_fp8_fp4_mega_moe
Align the registry name and module path with the upstream DeepGEMM
implementation file (sm100_fp8_fp4_mega_moe.cuh) and with the naming the
other SM100 deepgemm ports use.

The bench-suite config file is renamed too: run.py resolves a workload's
config by rglob("<kernel>.yaml"), so the file name must track the kernel
name. baseline.json/baseline.md are updated in place to keep regression
continuity for the two default workloads.

Verified: registry discovery, bench_suite --check-imports, and the
p1_tok2_h1024_i512_e2_k1_bm16 correctness config on B200.
