+++ tirx_kernels/bench_suite/config/deepgemm/sm100_fp8_fp4_mega_moe.yaml
@@ -1,4 +1,4 @@
-kernel: deepgemm_fp8_fp4_mega_moe
+kernel: sm100_fp8_fp4_mega_moe
 defaults: {num_gpus: 1, timer: megamoe}
 configs:
   - {config: t64_m64_h7168_i3072_e384_k6_g1, default: true}
